html/requirements.txt
Flask
Pillow
Flask-SQLAlchemy
Flask-Migrate
